feat：Extract FakeRefreshControl to a separate file
