+++ EssentialFeed/EssentialFeediOSTests/Views/FakeRefreshControl.swift
@@ -1,0 +1,22 @@
+//
+//  FakeRefreshControl.swift
+//  EssentialFeediOSTests
+//
+//  Created by nicaho on 2024/8/6.
+//
+
+import UIKit
+
+class FakeRefreshControl: UIRefreshControl {
+    private var _isRefreshing = false
+    
+    override var isRefreshing: Bool { _isRefreshing }
+    
+    override func beginRefreshing() {
+        _isRefreshing = true
+    }
+    
+    override func endRefreshing() {
+        _isRefreshing = false
+    }
+}
